requirements.txt
# Core dependencies for XGBoost trading system
xgboost>=1.7.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
pandas-ta>=0.3.14b
backtesting>=0.3.3
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
google-cloud-aiplatform>=1.38.1
imbalanced-learn>=0.10.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0